requirements.txt
streamlit 
pandas 
plotly 
folium 
streamlit-folium 
requests 
openpyxl 
matplotlib 
gspread 
google-auth 
